#12 - Camera is now working with GrabbableUI component
The UI elements (AnatomicViewer and Slides) need to be moved to appear on the screen, because their initial reference does not assume a scene node but a space position.
It would be best if we change the name of GrabbleUI to Orbitable or something else, since it is not restricted to UI elements anymore.